- converted answer and deviation history to ts - added missing types - combined enhanced with components
