+++ feeds/src/components/aggregatorVis/AggregatorVis.tsx
@@ -1,11 +1,32 @@
-import AggregatorVis from './AggregatorVis.component'
+import React from 'react'
 import { connect } from 'react-redux'
 import { AppState } from 'state'
-
+import { Vis } from '../vis'
+import { FeedConfig } from 'feeds'
 import {
   aggregatorSelectors,
   aggregatorOperations,
 } from 'state/ducks/aggregator'
+
+interface StateProps {
+  config: FeedConfig
+  latestOraclesState: any
+  latestAnswerTimestamp: any
+  latestAnswer: any
+  pendingAnswerId: any
+  oracleAnswers: any
+  oracleList: any
+  latestRequestTimestamp: any
+  minimumAnswers: any
+}
+
+interface DispatchProps {
+  fetchJobId: any
+}
+
+interface Props extends StateProps, DispatchProps {}
+
+const AggregatorVis: React.FC<Props> = props => <Vis {...props}></Vis>
 
 const mapStateToProps = (state: AppState) => ({
   latestOraclesState: aggregatorSelectors.latestOraclesState(state),
