+++ feeds/src/components/answerHistory/AnswerHistory.enhanced.ts
@@ -1,9 +1,9 @@
-import { compose } from 'recompose'
 import { connect } from 'react-redux'
 import AnswerHistory from './AnswerHistory.component'
+import { AppState } from 'state'
 
-const mapStateToProps = state => ({
+const mapStateToProps = (state: AppState) => ({
   answerHistory: state.aggregator.answerHistory,
 })
 
-export default compose(connect(mapStateToProps))(AnswerHistory)
+export default connect(mapStateToProps)(AnswerHistory)
