--- feeds/src/components/oracleTable/OracleTable.enhanced.ts
@@ -1,18 +1,0 @@
-import { connect } from 'react-redux'
-import OracleTable from './OracleTable.component'
-import {
-  aggregatorSelectors,
-  aggregatorOperations,
-} from 'state/ducks/aggregator'
-import { AppState } from 'state'
-
-const mapStateToProps = (state: AppState) => ({
-  ethGasPrice: state.aggregator.ethGasPrice,
-  latestOraclesState: aggregatorSelectors.latestOraclesState(state),
-})
-
-const mapDispatchToProps = {
-  fetchEthGasPrice: aggregatorOperations.fetchEthGasPrice,
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(OracleTable)
