+++ feeds/src/components/oracleTable/index.ts
@@ -1,1 +1,1 @@
-export { default as OracleTable } from './OracleTable.enhanced'
+export { default as OracleTable } from './OracleTable'
